src/process_image/requirements.txt
pillow>=10.0.0
pillow-heif>=0.13.0
boto3>=1.28.0
